Tries to solver issue #196
This is not meant to be accepted as it is as the fix for issue #196 but it's meant to open a discussion
on how it should be properly fixed.

The problem is that `lib/directory.coffe` gets loaded lazily for every folder currently expanded on the ui,
doing it so means it'st not possible to know if there is any subfolder or subfile at any level which is not ignored
and hence should be shown.

This commits does a deep first search on all subdirectories when enumerating a directory on `directory/getEntries`.
If there is at least one non ignored file on any child on any level the directory is not ignored.