Pass filesystem stats to directory/file instances
Developers monitoring the tree-view may need to access file-system stats
for an entry. Having the stats available on the instance saves an author
from making another filesystem call. For very large directories, this is
a non-trivial boost to performance.

Rather than pass a reference to the stat variable directly, a new object
is created to give to the Directory or File instance. This might benefit
garbage collection, although this is based on little more than theory.