Increase spec resiliency in tree-view-spec.coffee
Sometimes `treeView.currentlyOpening` doesn't contain the expected `Promise`,
thus causing failures like the one in https://circleci.com/gh/atom/atom/841.
It's possible that such `Promise` finishes before the test has the chance to
call `then` on it, thus causing it to be removed from that `Map` (see
https://github.com/atom/tree-view/blob/06bdfae/lib/tree-view.coffee#L223).

With this commit, we'll now simply wait until that `Promise` is removed from the
`currentlyOpening` `Map`, which should be more resilient.