Apply .tree-view class to root element to simplify focus handling
Existing key-bindings target the `.tree-view` element, but previously
this element wasn't the root of the tree view. Now that focus will be
handled by the workspace, making existing `.tree-view` bindings work
would have required us to transfer focus from the outer element to the
element with the `.tree-view` class. Moving this class to the root
introduces some risk of breaking themes, but I think it's worth keeping
the focus handling simple.