Fix other flaky test re: adding new file
Similar to 55e1b25bc5315d60143859606523fd55cbef5f96, we need to wait for
the tree view selection to *fully* update before checking the selected
item in the tree view.

As far as I can tell, when a file is added, the tree view is updated in
a few steps:

1. Add the new file to the directory
2. Select the new file

Prior to this change, we were waiting for Step 1 to happen, and then
asserting that Step 2 had also already happened. Between Step 1 and Step
2, it appears that the *directory* is selected, and that was causing the
failure seen in
https://github.com/atom/tree-view/issues/1155#issuecomment-319471558.